Jan 20th

+++ Scripts/ArmamentScripts/ArmamentStateMachine/AimedTorpedoStateMachine.cs
@@ -6,7 +6,7 @@ using AimedTorpedoStates;
 public class AimedTorpedoStateMachine : StateMachineBase
 {
     [HideInInspector]
-    public int reload_time;
+    public float reload_time;
 
     [HideInInspector]
     public AimedTorpedoTubeController controller;
